requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
datasets==2.21.0
dill==0.3.8
filelock==3.21.2
frozenlist==1.8.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.16
numpy==2.4.2
packaging==26.0
pandas==3.0.0
propcache==0.4.1
pyarrow==23.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
shellingham==1.5.4
six==1.17.0
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xxhash==3.6.0
yarl==1.22.0
